requirements.txt
torch==2.4.0
torchvision==0.19.0
pillow==10.4.0